requirements.txt
# Core Dependencies
torch>=1.10.0
torchvision>=0.11.0
numpy>=1.20.0
pillow>=8.0.0

# Data Processing
pyyaml>=5.4.0
scikit-learn>=0.24.0

# Visualization
matplotlib>=3.3.0
seaborn>=0.11.0

# Training & Logging
tqdm>=4.60.0
tensorboard>=2.8.0

# Optional (for UMAP visualizations)
umap-learn>=0.5.0
